# openapi.yaml
openapi: 3.0.0
info:
  title: AIQuiz API
  description: Complete API documentation for the AIQuiz educational platform with AI-powered question generation and RAG-enabled document processing
  version: 1.0.0
  contact:
    name: AIQuiz Team
servers:
  - url: /api
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/manager/auth/login
  schemas:
...
    User:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        faculty:
          type: string
        department:
          type: string
        role:
          type: string
          enum: [admin, professor]
        lastLogin:
          type: string
          format: date-time
...
paths:
...
  /manager/auth/login:
    post:
      summary: Manager authentication
      description: Authenticate manager and get JWT token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /manager/auth/validate-reset-token:
    post:
      summary: Validate password reset token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
      responses:
        '200':
          description: Token is valid
  /manager/auth/validate-invitation:
    post:
      summary: Validate invitation token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
      responses:
        '200':
          description: Token is valid
...
  /manager/rag/search:
    post:
      summary: Semantic search through documents
      description: Search through uploaded documents using RAG
      tags:
        - RAG Search
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
              properties:
                query:
                  type: string
                  description: Search query
                limit:
                  type: integer
                  default: 5
                  description: Number of results to return
                subjectId:
                  type: string
                  description: Filter by subject
                topicId:
                  type: string
                  description: Filter by topic
                subtopicId:
                  type: string
                  description: Filter by subtopic
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        content:
                          type: string
                        score:
                          type: number
                        metadata:
                          type: object

  # Quiz and Student Data
...